code/requirements.txt
# Function dependencies, for example:
# package>=version
oauth2client
google-cloud-storage
google-cloud-bigquery